Fix failure during running uber jar
Add the ballerina-java gradle module to rt jar to fix the failure during running the uber-jar
Resolve https://github.com/ballerina-platform/ballerina-lang/issues/26235
