+++ langlib/ballerina-java/src/main/java/module-info.java
@@ -1,4 +1,4 @@
 module io.ballerina.java {
     requires io.ballerina.jvm;
-    exports org.ballerinalang.nativeimpl.java;
+    exports org.ballerinalang.langlib.java;
 }
